requirements.txt
flask==2.0.1
Werkzeug==2.0.3
openai>=1.0.0
python-dotenv==0.19.0
requests==2.31.0
gunicorn 